% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxtoexact_func.R
\name{approxtoexact_func}
\alias{approxtoexact_func}
\title{Convert the approximate allocation (proportion) to exact allocation (integer) without constraint}
\usage{
approxtoexact_func(n, w)
}
\arguments{
\item{n}{Sample size, must be a positive integer}

\item{w}{Approximate allocation/proportion, must be a real-valued vector, can get from running liftone_constrained_GLM or liftone_constrained_MLM}
}
\value{
allocation is the exact allocation or integer value of the number of subjects sampled from the group
}
\description{
Convert the approximate allocation (proportion) to exact allocation (integer) without constraint
}
\examples{
exact_design = approxtoexact_func(n=600, w=c(0.2593526, 0.0000000, 0.0000000,
0.1565024, 0.2891565, 0.0000000, 0.0000000, 0.2949885))


}
