% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_uniform.R
\name{constrained_uniform}
\alias{constrained_uniform}
\title{Find (constrained) uniform exact allocation of the study for bounded design}
\usage{
constrained_uniform(Ni, nsample)
}
\arguments{
\item{Ni}{a vector with size m, upper bound for exact design of each category/stratification group, if unconstrained, use Inf vector, the function will return unbounded uniform allocation}

\item{nsample}{a number, the sample size}
}
\value{
n is the constrained/unconstrained uniform exact allocation
}
\description{
Find (constrained) uniform exact allocation of the study for bounded design
}
\examples{

constrained_uniform(Ni=c(50, 40, 10, 200, 150, 50), nsample=200)

}
