% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fdet_func_unif.R
\name{Fdet_func_unif}
\alias{Fdet_func_unif}
\title{Determinant function to be used for finding constrained uniform samplings}
\usage{
Fdet_func_unif(w, beta = NULL, X = NULL, link = NULL)
}
\arguments{
\item{w}{allocation (can be exact or approximate)}

\item{beta}{use NULL (default to be NULL)}

\item{X}{use NULL (default to be NULL)}

\item{link}{use NULL (default to be NULL)}
}
\value{
product of all allocation
}
\description{
Determinant function to be used for finding constrained uniform samplings
}
\examples{
Fdet_func_unif(w=c(0.2,0.2,0.2,0.2,0.2))


}
