% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iset_func_trauma.R
\name{iset_func_trauma}
\alias{iset_func_trauma}
\title{trauma_data example (see Huang, Tong, Yang (2023)) specific function for finding index set that if allocation of that index add "1", the new allocation still falls within the constraint
Used in approxtoexact_constrained_func()}
\usage{
iset_func_trauma(allocation)
}
\arguments{
\item{allocation}{the exact allocation}
}
\value{
list of TRUE and FALSE, if TRUE, it means the allocation of this index will fall out of the constraint with more subject; if TURE, it means the allocation of this index can add more subjects
}
\description{
trauma_data example (see Huang, Tong, Yang (2023)) specific function for finding index set that if allocation of that index add "1", the new allocation still falls within the constraint
Used in approxtoexact_constrained_func()
}
\examples{
iset_func_trauma(allocation=c(50,30,10,10,100,100,200,10))



}
