% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.matrix_list.R
\name{print.matrix_list}
\alias{print.matrix_list}
\title{Print Method for matrix_list Objects}
\usage{
\method{print}{matrix_list}(x, ...)
}
\arguments{
\item{x}{An object of class `matrix_list`.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns `x` (the input object).
}
\description{
Custom print method for objects of class `matrix_list`.
}
