% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.matrix_output.R
\name{print.matrix_output}
\alias{print.matrix_output}
\title{Print Method for matrix_output Objects}
\usage{
\method{print}{matrix_output}(x, ...)
}
\arguments{
\item{x}{An object of class `matrix_output`.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns `x` (the input object).
}
\description{
Custom print method for objects of class `matrix_output`.
}
